Some fixes. Now 'delete automat' is working.

+++ Casino/src/main/java/softgroup/ua/repository/AutomatRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Created by Вова on 04.03.2017.
  */
 @Repository
-public interface AutomatRepository extends JpaRepository <AutomatEntity, Long> {
+public interface AutomatRepository extends JpaRepository <AutomatEntity, Integer> {
     public List <AutomatEntity> findByAutomatId (int automatId);
     public List <AutomatEntity> findByAutomatName (String automatName);
    // public List <AutomatEntity> findAll();
